requirements.txt
aiohttp==3.8.4+computecanada
aiosignal==1.3.1+computecanada
anyio==3.6.2+computecanada
arff==0.9+computecanada
argon2_cffi==21.3.0+computecanada
argon2_cffi_bindings==21.2.0+computecanada
asttokens==2.2.1+computecanada
async_generator==1.10+computecanada
async_timeout==4.0.2+computecanada
attrs==22.2.0+computecanada
backcall==0.2.0+computecanada
backports-abc==0.5+computecanada
backports.shutil_get_terminal_size==1.0.0+computecanada
bcrypt==4.0.1+computecanada
beautifulsoup4==4.11.2+computecanada
bitstring==4.0.1+computecanada
bleach==6.0.0+computecanada
certifi==2022.12.7+computecanada
cffi==1.15.1+computecanada
chardet==5.1.0+computecanada
charset_normalizer==3.0.1+computecanada
comm==0.1.2+computecanada
contourpy==1.0.7+computecanada
cryptography==39.0.1+computecanada
cycler==0.11.0+computecanada
Cython==0.29.33+computecanada
datasets==2.13.0+computecanada
deap==1.3.3+computecanada
debugpy==1.6.6+computecanada
decorator==5.1.1+computecanada
defusedxml==0.7.1+computecanada
dill==0.3.6+computecanada
dnspython==2.3.0+computecanada
ecdsa==0.18.0+computecanada
entrypoints==0.4+computecanada
evaluate==0.4.0+computecanada
executing==1.2.0+computecanada
fastjsonschema==2.16.2+computecanada
filelock==3.12.0+computecanada
fonttools==4.38.0+computecanada
frozenlist==1.3.3+computecanada
fsspec==2023.6.0+computecanada
funcsigs==1.0.2+computecanada
huggingface_hub==0.15.1+computecanada
idna==3.4+computecanada
importlib_metadata==5.2.0+computecanada
importlib_resources==5.12.0+computecanada
ipykernel==6.21.2+computecanada
ipython==8.10.0+computecanada
ipython_genutils==0.2.0+computecanada
ipywidgets==8.0.4+computecanada
jedi==0.18.2+computecanada
Jinja2==3.1.2+computecanada
jsonschema==4.17.3+computecanada
jupyter_client==8.0.3+computecanada
jupyter_core==5.2.0+computecanada
jupyter_events==0.6.3+computecanada
jupyter_server==2.3.0+computecanada
jupyter_server_terminals==0.4.4+computecanada
jupyterlab_pygments==0.2.2+computecanada
jupyterlab_widgets==3.0.5+computecanada
kiwisolver==1.4.4+computecanada
lockfile==0.12.2+computecanada
MarkupSafe==2.1.2+computecanada
matplotlib==3.7.0+computecanada
matplotlib_inline==0.1.6+computecanada
mistune==2.0.5+computecanada
mock==5.0.1+computecanada
mpmath==1.2.1+computecanada
multidict==6.0.4+computecanada
multiprocess==0.70.14+computecanada
nbclassic==0.5.2+computecanada
nbclient==0.7.2+computecanada
nbconvert==7.2.9+computecanada
nbformat==5.7.3+computecanada
nest_asyncio==1.5.6+computecanada
netaddr==0.8.0+computecanada
netifaces==0.11.0+computecanada
networkx==3.1+computecanada
nose==1.3.7+computecanada
notebook==6.5.2+computecanada
notebook_shim==0.2.2+computecanada
numpy==1.24.2+computecanada
packaging==23.0+computecanada
pandas==1.5.3+computecanada
pandocfilters==1.5.0+computecanada
paramiko==3.0.0+computecanada
parso==0.8.3+computecanada
path==16.6.0+computecanada
path.py==12.5.0+computecanada
pathlib2==2.3.7+computecanada
paycheck==1.0.2+computecanada
pbr==5.11.1+computecanada
pexpect==4.8.0+computecanada
pickleshare==0.7.5+computecanada
Pillow==9.4.0+computecanada
pkgutil_resolve_name==1.3.10+computecanada
platformdirs==2.5.2+computecanada
prometheus_client==0.16.0+computecanada
prompt_toolkit==3.0.37+computecanada
psutil==5.9.4+computecanada
ptyprocess==0.7.0+computecanada
pure_eval==0.2.2+computecanada
pyarrow==11.0.0
pycparser==2.21+computecanada
Pygments==2.14.0+computecanada
PyNaCl==1.5.0+computecanada
pyparsing==3.0.9+computecanada
pyrsistent==0.19.3+computecanada
python-dateutil==2.8.2+computecanada
python_json_logger==2.0.7+computecanada
pytz==2022.7.1+computecanada
PyYAML==6.0+computecanada
pyzmq==25.0.0+computecanada
regex==2022.10.31+computecanada
requests==2.28.2+computecanada
responses==0.18.0+computecanada
rfc3339_validator==0.1.4+computecanada
rfc3986_validator==0.1.1+computecanada
safetensors==0.3.1+computecanada
scipy==1.10.1+computecanada
Send2Trash==1.8.0+computecanada
simplegeneric==0.8.1+computecanada
singledispatch==4.0.0+computecanada
six==1.16.0+computecanada
sniffio==1.3.0+computecanada
soupsieve==2.4+computecanada
stack_data==0.6.2+computecanada
sympy==1.11.1+computecanada
terminado==0.17.1+computecanada
testpath==0.6.0+computecanada
tinycss2==1.2.1+computecanada
tokenizers==0.13.3+computecanada
torch==2.0.1+computecanada
tornado==6.2+computecanada
tqdm==4.65.0+computecanada
traitlets==5.9.0+computecanada
transformers==4.30.2+computecanada
typing_extensions==4.5.0+computecanada
urllib3==1.26.14+computecanada
wcwidth==0.2.6+computecanada
webencodings==0.5.1+computecanada
websocket_client==1.5.1+computecanada
widgetsnbextension==4.0.5+computecanada
xxhash==3.2.0+computecanada
yarl==1.8.2+computecanada
zipp==3.14.0+computecanada

